refactor 소셜로그인과 일반 api 설정 분리
