+++ front/src/api/pet.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import instance from './axiosInstance';
+import instance from '../lib/axiosInstance';
 
 export const fetchPetInfo = async (matchId: number) => {
   const res = await instance.get(`/api/matches/${matchId}/pet`);
