+++ front/src/lib/socialAxios.ts
@@ -1,0 +1,8 @@
+import axios from 'axios';
+
+export const socialAxios = axios.create({
+  baseURL: process.env.NEXT_PUBLIC_BACKEND_ORIGIN,
+  withCredentials: true,
+});
+
+export default socialAxios;
